requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-docx==1.1.0
PyPDF2==3.0.1
nltk>=3.8
python-multipart==0.0.9
